src/test/java/search/PhoneDictonaryTest.java
package search;

import junit.framework.TestCase;
import org.junit.Test;

import java.util.ArrayList;
import java.util.List;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

public class PhoneDictonaryTest {
    @Test
    public void whenFindByName() {
        PhoneDictonary phones = new PhoneDictonary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        List<Person> persons = phones.find("Petr");
        assertThat(persons.get(0).getSurname(), is("Arsentev"));
    }

    @Test
    public void whenNoFindByName() {
        PhoneDictonary phones = new PhoneDictonary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        List<Person> persons = phones.find("Ivan");
        assertThat(persons.size(), is(0));
    }
}